fixup move tool to core
