+++ miso-web/src/main/resources/external/miso.properties
@@ -97,7 +97,7 @@ miso.pools.strictIndexChecking:false
 ## Customize warnings for duplicate or near-duplicate indices.
 # bcl2fastq cannot distinguish between indices with 0 or 1 mismatches,
 #   so default error edit distance is set to 1 (will also flag on 0 mismatches).
-miso.pools.error.index.mismatches:0
+miso.pools.error.index.mismatches:1
 miso.pools.error.index.mismatches.message:DUPLICATE INDICES
 # bcl2fastq can be run manually with setting "0 mismatches" when there are 2 mismatches between indices,
 #   so default warning edit distance is set to 2.
