+++ miso-web/src/main/webapp/scripts/indexDistance_ajax.js
@@ -36,11 +36,11 @@
         data: JSON.stringify(data)
       }).success(function(data) {
         var results = '';
-        if (data.warningEditDistance > data.errorEditDistance) {
-          results += 'Indices with edit distance of less than ' + Constants.warningEditDistance + ' are considered near-matches.\n';
+        if (Constants.warningEditDistance > Constants.errorEditDistance) {
+          results += 'Indices with ' + Constants.warningEditDistance + ' or fewer mismatches are considered near-matches.\n';
         }
-        results += 'Indices with edit distance of less than ' + Constants.errorEditDistance + ' are considered DUPLICATES.\n\n';
-        results += 'Indices were compared at the length of the shortest Index 1, which is ' + data.shortestIndexLength + ' bp.\n';
+        results += 'Indices with ' + Constants.errorEditDistance + ' or fewer mismatches are considered DUPLICATES.\n\n';
+        results += 'Indices were compared at the length of the shortest index, which is ' + data.shortestIndexLength + ' bp.\n\n';
         if (!data.collisions || !data.collisions.length) {
           results += 'All sequences are at least ' + minDistance + ' edits apart';
           $('#results').css('color', 'green');
